Adjust name and title sizes at various viewport widths.
I thought I left this project at a point where it was responding nicely, but having returned to it I saw that the header and title were getting larger too early and causing layout issues.

Is it possible Chrome updates would have changed the spacing of text in some way?