Revise design to use flexbox more fully.
This is a large check-in. Most of the page has been revised. First, to address the reviewers comments about mixing flexbox with "display: block". Then while reading more about flexbox, I was exposed to the concept of media objects, and then cards and I attempted to incorporate those ideas as well. Finally, the images were revised to both look better and to better reflect their corresponding project.

The current page is still rough and needs more work.